v6
V6 is going to be a huge updates with many new features and a completely revamp way of debloating your devices making use of our 'automation system' that has been in the works for a few weeks now. For now, everything is being kept private until it's public release. But here's whats going to change:
- The.. Logo! Ha